+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -10,8 +10,18 @@ however, it has to be formatted properly to pass verification tests.
 ## [0.9.4-preview] - 2099-1-1
 
 ### Fixed
-### Actions
+#### Actions
+
 ### Changed
+
+- `InputUser.onUnpairedDeviceUsed` now receives a 2nd argument which is the event that triggered the callback.
+  * Also, the callback is now triggered __BEFORE__ the given event is processed rather than after the event has already been written to the device. This allows updating the pairing state of the system before input is processed.
+  * In practice, this means that, for example, if the user switches from keyboard&mouse to gamepad, the initial input that triggered the switch will get picked up right away.
+
+#### Actions
+
+- When switching devices/controls on actions, the system will no longer subsequently force an initial state check on __all__ actions. Instead, every time an action's bindings get re-resolved, the system will simply cancel all on-going actions and then re-enable them the same way it would happen by manually calling `InputAction.Enable`.
+
 ### Added
 
 ## [0.9.3-preview] - 2019-8-15